refreshToken&AuthGuard
Created at : 1/09/2565
Update at : 4/09/2565
Author : Atsawin Anantasri
Detail :
- add refresh token when user idle (in this commit 55second to idle status) website will auto logout if user using website token will refresh

+++ src/app/auth.service.ts
@@ -10,24 +10,49 @@ import { LocalizedString } from '@angular/compiler';
 export class AuthService {
   constructor(private router: Router, private httpClient: HttpClient) {}
 
+  userInactivate:boolean = false
+
+  reFreshToken() {
+    return this.httpClient.post('http://127.0.0.1:8000/api/auth/refresh', {
+      token: localStorage.getItem('tokenLocal'),
+    });
+  }
+
   private tokenExpired(token: string) {
     const expiry = JSON.parse(atob(token.split('.')[1])).exp;
     return Math.floor(new Date().getTime() / 1000) >= expiry;
   }
 
-  CheckTokenTimeOut(): boolean {
+  async CheckTokenTimeOut(): Promise<boolean> {
     if (this.tokenExpired(localStorage.getItem('tokenLocal')!)) {
-      localStorage.clear()
+      if(this.userInactivate)
+      {
+        localStorage.clear();
+        console.log('refresh userInactivate !')
+      }
+      else
+      {
+        this.reFreshToken().subscribe({
+          next:(res:any) =>
+          {
+            console.log('token refresh success')
+            localStorage.setItem('tokenLocal',res.data.original.access_token)
+          },
+          error(err) {
+          },
+        })
+      }
       return true;
     } else {
       return false;
     }
   }
 
-  isLoggedin() {
-    this.CheckTokenTimeOut();
+    isLoggedin() {
+    this.CheckTokenTimeOut()
+    console.log('user activate!')
     const token = localStorage.getItem('tokenLocal');
-    console.log('isLoggedin',token != null)
+    console.log('isLoggedin', token != null);
     return token != null;
   }
 
@@ -35,5 +60,4 @@ export class AuthService {
     localStorage.clear();
     this.router.navigate(['']);
   }
-  
 }
